backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi
uvicorn[standard]
pytest
pytest-cov
httpx
python-dotenv

av==16.1.0
certifi==2026.1.4
click==8.3.1
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
paho-mqtt==2.1.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.3.32
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
tzdata==2025.3